refactor(seahaven-file): reorganize content module
- Moved the `Content` struct and its associated serialization and deserialization logic into separate modules (`ser.rs` and `de.rs`) for better organization.
- Introduced a new `root.rs` file to define the `Content` struct, improving the structure of the content model.
- This refactor enhances maintainability and readability of the codebase.
